Flag test as xfail
Avoid the issue described in this issue:
https://github.com/johbo/pip2nix/issues/35

During the test run in a nix-build we see calls to nix again which fail. Will
keep the issue open until there is a proper solution available.

Thanks to https://github.com/Riptawr for finding this problem and raising the
issue on Github.