Use packageOverrides to inject overlays
Due to changes in how the pythonPackages are exposed, it seems not to be
possible anymore to easily "unfix" them and get the "pythonPackagesFun" out of
it again.

packageOverrides is able to inject one overlay when the various layers are
composed into the package set.

Via composeExtensions the local overlays are first combined into a single one
and then passed as one overlay into packageOverrides.